Fix constructor being converted to undefined
The issue was that regular property accessing in javascript can inherit from the prototype. By default, each object in js has the "object" prototype which has a "constructor" property on it (Which is a function). I fixed this by createing the character Map so that it doesn't inherit from the default object.